requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
asgiref==3.6.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
beautifulsoup4==4.12.1 ; python_full_version >= '3.6.0'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
crispy-bootstrap5==0.7
cssselect==1.2.0 ; python_version >= '3.7'
dj-rest-auth==3.0.0
django==4.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==23.1
djangorestframework==3.14.0
feedfinder2==0.0.4
feedparser==6.0.10 ; python_version >= '3.6'
filelock==3.10.7 ; python_version >= '3.7'
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
jieba3k==0.35.1
joblib==1.2.0 ; python_version >= '3.7'
lxml==4.9.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
multidict==6.0.4 ; python_version >= '3.7'
newspaper3k==0.2.8
nltk==3.8.1 ; python_version >= '3.7'
openai==0.27.4
pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
pyyaml==6.0 ; python_version >= '3.6'
regex==2023.3.23 ; python_version >= '3.8'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.4 ; python_version >= '3.7'
sqlparse==0.4.3 ; python_version >= '3.5'
tinysegmenter==0.3
tldextract==3.4.0 ; python_version >= '3.7'
tqdm==4.65.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
yarl==1.8.2 ; python_version >= '3.7'
